pyproject.toml
[tool.poetry]
name = "telegram-bot-template"
version = "0.1.0"
description = "Template for telegram bot with payment system"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^3.16.0"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"
loguru = "^0.7.3"
telethon = "^1.38.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
